Move tests from CompactionJobRunTest to core module
